Refactoring file structure and style changes
