+++ README.md
@@ -4,8 +4,9 @@ SPDX-FileCopyrightText: 2022 The Aalto Grades Developers
 SPDX-License-Identifier: MIT
 -->
 
-# base-repository
-Repo for Aalto-Grades Program
+# Aalto Grades
+
+Repo for Aalto Grades Program
 
 ## Development build
