+++ server/.eslintrc.json
@@ -34,6 +34,33 @@
       "error",
       "always"
     ],
+    "max-len": [
+      "error",
+      {
+        "code": 100
+      }
+    ],
+    "brace-style": [
+      "error",
+      "1tbs",
+      {
+        "allowSingleLine": false
+      }
+    ],
+    "space-before-blocks": [
+      "error",
+      "always"
+    ],
+    "space-before-function-paren": [
+      "error",
+      {
+        "anonymous": "always",
+        "named": "never"
+      }
+    ],
+    "no-trailing-spaces": [
+      "error"
+    ],
     "@typescript-eslint/typedef": [
       "error",
       {
