+++ server/mockData/course_translations.sql
