--- server/src/configs/config.ts
@@ -1,7 +1,0 @@
-// SPDX-FileCopyrightText: 2023 The Aalto Grades Developers
-//
-// SPDX-License-Identifier: MIT
-
-export const axiosTimeout: number = 5000;
-export const jwtExpirySeconds: number = 25 * 60 * 60; // 25 hours
-export const jwtCookieExpiryMs: number = 24 * 60 * 60 * 1000; // 24 hours
