+++ server/src/configs/constants.ts
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: 2023 The Aalto Grades Developers
+//
+// SPDX-License-Identifier: MIT
+
+// This file contains all hard coded constants.
+
+export const AXIOS_TIMEOUT: number = 5000;
+
+// 25 hours in seconds.
+export const JWT_EXPIRY_SECONDS: number = 25 * 60 * 60;
+
+// 24 hours in milliseconds.
+export const JWT_COOKIE_EXPIRY_MS: number = 24 * 60 * 60 * 1000;
