+++ server/src/configs/environment.ts
@@ -2,19 +2,28 @@
 //
 // SPDX-License-Identifier: MIT
 
-const SISU_API_KEY: string | undefined = process.env.SISU_API_KEY;
-const SISU_API_URL: string | undefined = process.env.SISU_API_URL;
-const NODE_ENV: string = process.env.NODE_ENV ?? 'development';
+// This file reads all environment variables and defines their values as constants.
+
 const parsedPort: number = Number(process.env.AALTO_GRADES_BACKEND_PORT);
-const PORT: number = isNaN(parsedPort) ? 3000 : parsedPort;
+export const PORT: number = isNaN(parsedPort) ? 3000 : parsedPort;
+
+export const SISU_API_KEY: string | undefined = process.env.SISU_API_KEY;
+export const SISU_API_URL: string | undefined = process.env.SISU_API_URL;
+export const NODE_ENV: string = process.env.NODE_ENV ?? 'development';
 
 if ((!SISU_API_KEY || !SISU_API_URL) && NODE_ENV === 'production') {
-  throw new Error('SISU_API_KEY and/or SISU_API_URL environment variable(s) undefined, required in production');
+  throw new Error(
+    'SISU_API_KEY and/or SISU_API_URL environment variable(s) undefined, required in production.'
+  );
 }
 
-export {
-  SISU_API_KEY,
-  SISU_API_URL,
-  NODE_ENV,
-  PORT
-};
+export const TEST_ENV: boolean = process.env.AALTO_GRADES_TEST_ENVIRONMENT === 'true';
+export const JWT_SECRET: string = process.env.AALTO_GRADES_JWT_SECRET || 'TOP_SECRET';
+export const FRONTEND_ORIGIN: string =
+  process.env.AALTO_GRADES_FRONTED_CORS_ORIGIN || 'http://localhost:3005';
+
+if (JWT_SECRET === 'TOP_SECRET') {
+  console.warn(
+    'No AALTO_GRADES_JWT_SECRET specified, using default value. Do not do this in production.'
+  );
+}
