+++ server/src/database/models/courseInstance.ts
@@ -2,12 +2,18 @@
 //
 // SPDX-License-Identifier: MIT
 
-import { DataTypes, Model, InferAttributes, InferCreationAttributes, CreationOptional, ForeignKey } from 'sequelize';
+import {
+  CreationOptional, DataTypes, ForeignKey, Model, InferAttributes, InferCreationAttributes
+} from 'sequelize';
+
 import { sequelize } from '..';
 import Course from './course';
 import User from './user';
 
-export default class CourseInstance extends Model<InferAttributes<CourseInstance>, InferCreationAttributes<CourseInstance>> {
+export default class CourseInstance extends Model<
+  InferAttributes<CourseInstance>,
+  InferCreationAttributes<CourseInstance>
+> {
   declare id: CreationOptional<number>;
   declare courseId: ForeignKey<Course['id']>;
   declare gradingType: string;
