--- server/src/services/index.ts
@@ -1,7 +1,0 @@
-// SPDX-FileCopyrightText: 2023 The Aalto Grades Developers
-//
-// SPDX-License-Identifier: MIT
-
-export * as courseService from './course';
-export * as userService from './user';
-export * as instanceService from './instance';
