+++ server/src/types/user.ts
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: 2023 The Aalto Grades Developers
+//
+// SPDX-License-Identifier: MIT
+
+export enum UserRole {
+  Student = 'STUDENT',
+  Teacher = 'TEACHER',
+  Admin = 'SYSADMIN',
+  Assistant = 'ASSISTANT',
+}
