+++ server/test/controllers/course.test.ts
@@ -2,92 +2,126 @@
 //
 // SPDX-License-Identifier: MIT
 
-import * as yup from 'yup';
-import { localizedStringSchema } from '../../src/controllers/course';
+import supertest from 'supertest';
 
-describe('Test localizedStringSchema', () => {
+import { app } from '../../src/app';
 
-  const runValidation = async (input: object): Promise<void> => {
-    const schema: yup.AnyObjectSchema = yup.object().shape({
-      name: localizedStringSchema.required(),
-      department: localizedStringSchema
+const request: supertest.SuperTest<supertest.Test> = supertest(app);
+
+describe('Test GET /v1/courses/:courseId', () => {
+  it('should respond with correct data when course exists', async () => {
+    const res: supertest.Response = await request.get('/v1/courses/1');
+    expect(res.body.success).toBe(true);
+    expect(res.body.course).toBeDefined();
+    expect(res.body.error).not.toBeDefined();
+    expect(res.body.course.id).toBe(1);
+    expect(res.body.course.courseCode).toBeDefined();
+    expect(res.body.course.department).toBeDefined();
+    expect(res.body.course.name).toBeDefined();
+    expect(res.body.course.evaluationInformation).toBeDefined();
+    expect(res.statusCode).toBe(200);
+  });
+
+  it('should respond with 404 not found, if non-existing course id', async () => {
+    const res: supertest.Response = await request.get('/v1/courses/-1');
+    expect(res.body.success).toBe(false);
+    expect(res.body.course).not.toBeDefined();
+    expect(res.body.error).toBeDefined();
+    expect(res.statusCode).toBe(404);
+  });
+
+  it('should respond with 400 bad request, if validation fails (non-number course id)',
+    async () => {
+      const res: supertest.Response = await request.get('/v1/courses/abc');
+      expect(res.body.success).toBe(false);
+      expect(res.body.course).not.toBeDefined();
+      expect(res.body.error).toBeDefined();
+      expect(res.statusCode).toBe(400);
     });
-    await schema.validate(input, { abortEarly: false });
-  };
+});
 
-  it('should resolve with correct input', async() => {
-    await expect(runValidation({
-      name: {
-        fi: 'Differentiaali- ja intergraalilaskenta 1',
-        en: 'Differential and Integral Calculus 1',
-        sv: 'Differential- och integralkalkyl 1'
-      },
-      department: {
-        fi: 'Matematiikan ja systeemianalyysin laitos',
-        en: 'Department of Mathematics and Systems Analysis',
-        sv: 'Institutionen för matematik och systemanalys'
-      }
-    })).resolves;
+describe('Test POST /v1/courses', () => {
 
-    await expect(runValidation({
-      name: {
-        fi: 'Differentiaali- ja intergraalilaskenta 1',
-        sv: 'Differential- och integralkalkyl 1'
-      },
-      department: {
-        fi: 'Matematiikan ja systeemianalyysin laitos',
-        en: 'Department of Mathematics and Systems Analysis'
-      }
-    })).resolves;
-  });
-
-  it('should resolve, if a non-required localized string is undefined', async() => {
-    await expect(runValidation({
-      name: {
-        fi: 'Differentiaali- ja intergraalilaskenta 1',
-        en: 'Differential and Integral Calculus 1',
-        sv: 'Differential- och integralkalkyl 1'
-      }
-    })).resolves;
-  });
-
-  it('should reject and throw ValidationError, if a required localized string is undefined', async () => {
-    await expect(runValidation({
-      department: {
-        fi: 'Matematiikan ja systeemianalyysin laitos',
-        en: 'Department of Mathematics and Systems Analysis',
-        sv: 'Institutionen för matematik och systemanalys'
-      }
-    }))
-      .rejects
-      .toThrow(yup.ValidationError);
-  });
-
-  it('should reject and throw ValidationError, if the object for a localized string is empty', async () => {
-    await expect(runValidation({
-      name: {
-        fi: 'Differentiaali- ja intergraalilaskenta 1',
-        en: 'Differential and Integral Calculus 1',
-        sv: 'Differential- och integralkalkyl 1'
-      },
-      department: {}
-    }))
-      .rejects
-      .toThrow(yup.ValidationError);
-  });
-
-  it('should reject and throw ValidationError, if unknown translations are present in a localized string', async () => {
-    await expect(runValidation({
-      name: {
-        test: 'not valid'
-      },
+  it('should respond with course data on correct input', async () => {
+    let input: object = {
+      courseCode: 'ELEC-A7200',
       department: {
         fi: 'Sähkötekniikan korkeakoulu',
-        test: 'not valid'
+        en: 'School of Electrical Engineering',
+        sv: 'Högskolan för elektroteknik'
+      },
+      name: {
+        fi: 'Signaalit ja järjestelmät',
+        en: 'Signals and Systems',
+        sv: ''
       }
-    }))
-      .rejects
-      .toThrow(yup.ValidationError);
+    };
+    let res: supertest.Response = await request.post('/v1/courses').send(input);
+    expect(res.statusCode).toBe(200);
+    expect(res.body.success).toBe(true);
+    expect(res.body.course.courseCode).toBe('ELEC-A7200');
+    expect(res.body.course.name).toStrictEqual({
+      'fi': 'Signaalit ja järjestelmät',
+      'en': 'Signals and Systems',
+      'sv': ''
+    });
+    expect(res.body.course.department).toStrictEqual({
+      'fi': 'Sähkötekniikan korkeakoulu',
+      'en': 'School of Electrical Engineering',
+      'sv': 'Högskolan för elektroteknik'
+    });
+
+    input = {
+      courseCode: 'ELEC-A7200',
+      department: {
+        fi: 'Sähkötekniikan korkeakoulu',
+        en: 'School of Electrical Engineering',
+      },
+      name: {
+        fi: 'Signaalit ja järjestelmät',
+        en: 'Signals and Systems',
+      }
+    };
+    res = await request.post('/v1/courses').send(input);
+    expect(res.statusCode).toBe(200);
+    expect(res.body.success).toBe(true);
+    expect(res.body.course.courseCode).toBe('ELEC-A7200');
+    expect(res.body.course.name).toStrictEqual({
+      'fi': 'Signaalit ja järjestelmät',
+      'en': 'Signals and Systems',
+      'sv': ''
+    });
+    expect(res.body.course.department).toStrictEqual({
+      'fi': 'Sähkötekniikan korkeakoulu',
+      'en': 'School of Electrical Engineering',
+      'sv': ''
+    });
   });
-    
+
+  it('should respond with validation errors, if required fields are undefined', async () => {
+    const input: object = {};
+    const res: supertest.Response = await request.post('/v1/courses').send(input);
+    expect(res.statusCode).toBe(400);
+    expect(res.body.success).toBe(false);
+    expect(res.body.errors).toContain('courseCode is a required field');
+    expect(res.body.errors).toContain('department is a required field');
+    expect(res.body.errors).toContain('name is a required field');
+  });
+
+  /* TODO: move next test case elsewhere in future, after refactoring commonly
+   * reusable functionality (e.g. middleware) to their own modules / functions
+   */
+
+  it('should respond with syntax error, if parsing request JSON fails', async () => {
+    const input: string = '{"courseCode": "ELEC-A7200"';
+    const res: supertest.Response = await request
+      .post('/v1/courses')
+      .send(input)
+      .set('Content-Type', 'application/json');
+    expect(res.statusCode).toBe(400);
+    expect(res.body.success).toBe(false);
+    expect(res.body.errors).toContain(
+      'SyntaxError: Unexpected end of JSON input: {"courseCode": "ELEC-A7200"'
+    );
+  });
 });
