+++ server/test/controllers/user.test.ts
@@ -1,0 +1,76 @@
+// SPDX-FileCopyrightText: 2023 The Aalto Grades Developers
+//
+// SPDX-License-Identifier: MIT
+
+import supertest from 'supertest';
+
+import { app } from '../../src/app';
+import { CoursesOfUser } from '../../src/controllers/user';
+
+const request: supertest.SuperTest<supertest.Test> = supertest(app);
+
+describe('Test GET /v1/user/:userId/courses', () => {
+  jest
+    .spyOn(global.Date, 'now')
+    .mockImplementation((): number => {
+      return new Date('2022-12-10').valueOf();
+    });
+
+  it('should respond with correct data', async () => {
+    const res: supertest.Response = await request.get('/v1/user/1/courses');
+    expect(res.body.success).toBe(true);
+    expect(res.statusCode).toBe(200);
+
+    expect(res.body.courses).toStrictEqual({
+      'current': [],
+      'previous': [
+        {
+          'id': 1,
+          'courseCode': 'CS-A1110',
+          'department': {
+            'fi': 'Tietotekniikan laitos',
+            'sv': 'Institutionen för datateknik',
+            'en': 'Department of Computer Science'
+          },
+          'name': {
+            'fi': 'Ohjelmointi 1',
+            'sv': 'Programmering 1',
+            'en': 'Programming 1'
+          },
+          'evaluationInformation': {
+            'fi': '',
+            'sv': '',
+            'en': ''
+          }
+        }
+      ]
+    });
+  });
+
+  it('should not contain duplicate courses', async () => {
+    const res: supertest.Response = await request.get('/v1/user/2/courses');
+    expect(res.body.success).toBe(true);
+    expect(res.statusCode).toBe(200);
+
+    const courses: CoursesOfUser = res.body.courses as CoursesOfUser;
+    let n: number = 0;
+    for (const i in courses.current) {
+      if (courses.current[i].courseCode === 'CS-A1120')
+        n++;
+    }
+
+    expect(n).toBe(1);
+  });
+
+  it('should correctly identify whether a course is current or previous', async () => {
+    const res: supertest.Response = await request.get('/v1/user/3/courses');
+    expect(res.body.success).toBe(true);
+    expect(res.statusCode).toBe(200);
+
+    const courses: CoursesOfUser = res.body.courses as CoursesOfUser;
+    expect(courses.current.length).toBe(1);
+    expect(courses.current[0].courseCode).toBe('PHYS-A1130');
+    expect(courses.previous.length).toBe(1);
+    expect(courses.previous[0].courseCode).toBe('TU-A1100');
+  });
+});
